MAJOR UPDATE
BUG FIXES:
No longer randomly resets the score while playing a game.
User is prompted with a warning to telling them all data will be lost if
they go back to the menu.

FEATURES:
Able to choose how ranking is displayed.
Able for the user to set the score for the players.
Able to see the diagram from wich the score is calculated (This still
needs a feature to zoom in on the diagram).